requirements.txt
pytest==9.0.2
selenium==4.41.0
allure-pytest==2.15.3
colorlog==6.10.1
pytest-xdist==3.8.0
PyYAML==6.0.3
allure-python-commons
openpyxl==3.1.5
pytest-rerunfailures
mysql-connector-python
requests
fastapi
uvicorn